--- msg/Collision.msg
@@ -1,6 +1,0 @@
-bool in_danger
-uint8 scan_idx
-float32 distance
-float32 wall_ang
-float32 speed
-bool copilot_active
